# docs/api/openapi.v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Concordat.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/environments/{env}/notifications": {
      "get": {
        "tags": [
          "Notifications"
        ],
        "summary": "List an environment's notification subscriptions",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notifications"
        ],
        "summary": "Subscribe a channel to an environment's events",
        "description": "An empty event list means every event. The alternative is a subscription that is configured, enabled, and silently delivers nothing — which looks correct from every screen. A WEBHOOK subscription also gets a signing secret, returned here once: every delivery to it carries an X-Concordat-Signature header, an HMAC-SHA256 of the body keyed on that secret, so the receiver can tell this registry sent it from anything else that found the URL.",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubscriptionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedSubscriptionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/environments/{env}/notifications/{id}/enabled": {
      "put": {
        "tags": [
          "Notifications"
        ],
        "summary": "Mute or unmute a subscription without losing its settings",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSubscriptionEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/environments/{env}/notifications/{id}": {
      "delete": {
        "tags": [
          "Notifications"
        ],
        "summary": "Delete a subscription",
        "parameters": [
          {
            "name": "env",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatedSubscriptionResponse": {
        "required": [
          "subscription",
          "signingSecret"
        ],
        "type": "object",
        "properties": {
          "subscription": {
            "description": "The subscription.",
            "$ref": "#/components/schemas/SubscriptionResponse"
          },
          "signingSecret": {
            "type": [
              "null",
              "string"
            ],
            "description": "The secret to configure on the receiving end, for a WEBHOOK subscription. <b>This is the\nonly time it is ever returned</b> — the registry stores only an encrypted reference and\ncannot show it again. Null for an EMAIL subscription, which has nothing to sign."
          }
        },
        "description": "A newly created subscription and its one-time webhook signing secret."
      },
...
      "CreateSubscriptionRequest": {
        "required": [
          "channel",
          "endpoint"
        ],
        "type": "object",
        "properties": {
          "channel": {
            "type": "string",
            "description": "`EMAIL` or `WEBHOOK`."
          },
          "endpoint": {
            "type": "string",
            "description": "An email address, or an `https` URL."
          },
          "events": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "type": "string"
            },
            "description": "Which events, or omit for all of them."
          }
        },
        "description": "Request to create a subscription."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SetSubscriptionEnabledRequest": {
        "required": [
          "enabled"
        ],
        "type": "object",
        "properties": {
          "enabled": {
            "type": "boolean",
            "description": "Whether to deliver."
          }
        },
        "description": "Request to mute or unmute a subscription."
      },
...
      "SubscriptionResponse": {
        "required": [
          "id",
          "channel",
          "endpoint",
          "events",
          "enabled",
          "createdAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Its identity.",
            "format": "uuid"
          },
          "channel": {
            "type": "string",
            "description": "`EMAIL` or `WEBHOOK`."
          },
          "endpoint": {
            "type": "string",
            "description": "Where notifications go."
          },
          "events": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Which events, or empty for all of them."
          },
          "enabled": {
            "type": "boolean",
            "description": "Whether it is currently delivering."
          },
          "createdAt": {
            "type": "string",
            "description": "When it was created.",
            "format": "date-time"
          }
        },
        "description": "A notification subscription."
      },
...
    }
  },
...
}